% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{afan}
\alias{afan}
\title{Converts class labels to the lowest set of possible integers.}
\usage{
afan(x)
}
\arguments{
\item{x}{A vector of class labels.}
}
\value{
A new (or identical) vector of class lables.
}
\description{
Converts class labels to the lowest set of possible integers.
}
\details{
If the largest class label in \code{x} is equal to the number of groups and the smallest group label is equal to 1, the class vector returned by \code{afan} will be identical.  If however, potential lower class labels are skiped, a new vector of class labels is returned.  This vector will ensure that no potential class labels have been skipped.
}
\examples{
afan(c(4,4,8,1,1,1,3,3,6))
afan(c(4,4,1,1,1,3,3,2,2))
}

