% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ajc69.R
\name{ajc69}
\alias{ajc69}
\title{Adjusted Jukes and Cantor 1969 Pairwise Distances}
\usage{
ajc69(DNAbin.data)
}
\arguments{
\item{DNAbin.data}{Aligned DNA sequence data in \code{DNAbin} (matrix) form.}
}
\value{
A pairwise distance matrix
}
\description{
Adjusted Jukes and Cantor 1969 Pairwise Distances
}
\details{
This function computes a matrix of pairwise distances from aligned DNA sequences using an ajusted version of the Jukes and Cantor 1969 model of DNA evolution.  This adjustment takes into account ambiguous IUPAC codes. For details on these calculations see the Supplmentary Information in Vrbik et al.
}
\examples{
library(ape)
data(woodmouse)
ajc69(woodmouse)
}
\references{
Irene Vrbik, David A Stephens, Michel Roger and Bluma G Brenner. The Gap Procedure: for the identification of phylogenetic clusters in HIV-1 sequence data.

Jukes, TH, Cantor, CR. \emph{Evolution of protein molecules} (1969)
}

