% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ak80.R
\name{ak80}
\alias{ak80}
\title{Adjusted Kimura 1980 Pairwise Distances}
\usage{
ak80(DNAbin.data)
}
\arguments{
\item{DNAbin.data}{Aligned DNA sequence data in \code{DNAbin} (matrix) form.}
}
\value{
A pairwise distance matrix
}
\description{
Adjusted Kimura 1980 Pairwise Distances
}
\details{
This function computes a matrix of pairwise distances from aligned DNA sequences using an ajusted version of the Kimura 1980 model of DNA evolution.  This adjustment takes into account ambiguous IUPAC codes. For details on these calculations see the Supplmentary Information in Vrbik et al.
}
\examples{
library(ape)
data(woodmouse)
ak80(woodmouse)
}
\references{
Irene Vrbik, David A Stephens, Michel Roger and Bluma G Brenner. The Gap Procedure: for the identification of phylogenetic clusters in HIV-1 sequence data.

Kiumra, M. \emph{A simple method for estimating evolutionary rates of base substitutions through comparative studies of nucleotide sequences}. Journal of molecular evolution \strong{16}(2), 111-120 (1980).
}

