% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asNumeric.R
\name{asNumeric}
\alias{asNumeric}
\title{Maps DNAbin class to the numbers 0 through 16.}
\usage{
asNumeric(DNAbin.data)
}
\arguments{
\item{DNAbin.data}{DNA sequence data in \code{DNAbin} (matrix) form.}
}
\value{
A matrix (or list) of DNA sequences with code values of 0--16
}
\description{
Maps DNAbin class to the numbers 0 through 16.
}
\details{
The DNAbin class is mapped to the number 0 through 16.  Below is the scheme
}

