% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{bit4Count}
\alias{bit4Count}
\title{Counts the number of bits in the A G C T position in the \code{DNAbin} bit-level coding scheme}
\usage{
bit4Count(xL, shift = TRUE)
}
\arguments{
\item{xL}{A DNA sequence in integer form}

\item{shift}{A logical indicating whether an initial shift needs to be made}
}
\value{
A vector of integer values indicating the number of bits in position A G C T at each site of \code{xL}
}
\description{
Counts the number of bits in the A G C T position in the \code{DNAbin} bit-level coding scheme
}
\examples{
# IUPAC code "A" has Bit-level code 1000 1000 which corresponds to the number 136
# there is 1 bit in the first four positions (corresponding to the binary coding system for A G C T  in the DNAbin class)
bit4Count(136)

# IUPAC code "B" (c, g or t) has Bit-level code 0111 0000 which corresponds to the number 112
# there are 3 bits in the first four positions (corresponding to the binary coding system for A G C T  in the DNAbin class)
bit4Count(112)
}

