% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{bitMatch}
\alias{bitMatch}
\title{Counts the number of posible pairwise matches between \code{xL} and \code{yL},}
\usage{
bitMatch(xL, yL)
}
\arguments{
\item{xL}{A DNA sequence in integer form.}

\item{yL}{A DNA sequence in integer form.}
}
\value{
A vector of integer values indicating the number of matching character comparisons between the IUPAC symbols \code{xL} and \code{yL}.
}
\description{
Counts the number of posible pairwise matches between \code{xL} and \code{yL},
}
\examples{
# IUPAC code "V" (a, c or g) has Bit-level code 1110 0000 which corresponds to the number 224
# IUPAC code "H" (a, c or t) has Bit-level code 1011 0000 which corresponds to the number 176
# eg. V (a, c or g) and H (a, c or t) has 9 pairwise comparisons aa, ac, at, ca, cc, ct, ga, gc, gt
# and has 2 matching comparison (aa, cc).
bitMatch(224,176)
}

