% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxplotWB.R
\name{boxplotWB}
\alias{boxplotWB}
\title{Plots the side-by-side boxplot of the within-cluster and between-cluster pairwise distances.}
\usage{
boxplotWB(GPout, classvec = NULL, Gindex = NULL, hcutoff = 0.05,
  show.singletons = FALSE, savefile = NULL, ...)
}
\arguments{
\item{GPout}{The output object created by the \code{GapProcedure} function.}

\item{classvec}{A vector indicating the class memberships of the data used to produce \code{GPout}.  The default (NULL) takes the classification vector stored in \code{GPout}.}

\item{Gindex}{A vector of cluster indices. If set to NULL (default), then the entire set of class labels is considered.}

\item{savefile}{A character vector.  If set to \code{NULL} (default), the figures are not saved and ploted in R, otherwise the figure will be saved to \code{<savefile>Cluster<g>.pdf}, where \code{g} denotes the cluster number.}

\item{...}{options for \code{pdf()}}
}
\description{
Plots the side-by-side boxplot of the within-cluster and between-cluster pairwise distances.
}
\details{
For each class label (\eqn{g = 1, \dots, G}), a side-by-side boxplot is created.
}
\seealso{
\code{\link[grDevices]{pdf}}
}

