% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{diffpossitive}
\alias{diffpossitive}
\title{Sorts a vector and computers the difference between sucessive numbers}
\usage{
diffpossitive(x, needtosort = FALSE, value = FALSE)
}
\arguments{
\item{x}{A vector of non-negative values.}
}
\value{
An integer indicating where largest difference in the sorted \code{x} vector is obsvered.
}
\description{
Sorts a vector and computers the difference between sucessive numbers
}
\details{
Note that (1) comparisons with 0 are not considered for the gap search (2) in the presence of a tie, the first large gap will be returned (see examples).
}
\examples{
# Consider this vector of sorted values
x <- c(0,3,6.1,7,8)

# The largest gap in x occurs between the second and third value.
diffpossitive(x)

# This value should not change after we've shuffled x
sx <- sample(x)
diffpossitive(x)
cat("The gap occurs between sorted position", diffpossitive(x),"and", diffpossitive(x)+1)

# Note that ties will return the location of the first gap
diffpossitive(c(1,2,4,8,9,10,14))
cat("The gap occurs between sorted position", diffpossitive(x),"and", diffpossitive(x)+1)
}

