% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractClusterMembers.R
\name{extractClusterMembers}
\alias{extractClusterMembers}
\title{Extract clusters of a particular size}
\usage{
extractClusterMembers(classification.vector, small = NULL, big = NULL)
}
\arguments{
\item{classification.vector}{A classification vector.}

\item{small}{Specifies the minimum number of members the cluster contain}

\item{big}{The minimum number of members the cluster must contain}
}
\value{
groups A vector of class labels that satisfy the requirements described in \code{Details}.

ind A vector of indices indicating which elements of \code{classification.vector} belong to one of the clusters stored in \code{groups}.
}
\description{
Extract clusters of a particular size
}
\details{
This function returns the cluster labels (and members) which have between \code{small} and \code{big} members
}

