% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GapPlot.R
\name{GapPlot}
\alias{GapPlot}
\title{Plots the sorted pairwise distances and indicates the position of the largest gap.}
\usage{
GapPlot(GPout, classvec = NULL, Nindex = NULL, hcutoff = NULL,
  show.singletons = TRUE, savefile = NULL, ...)
}
\arguments{
\item{GPout}{The output object created by the \code{GapProcedure} function.}

\item{classvec}{A vector indicating the class memberships of the data used to produce \code{GPout}.  The default (NULL) takes the classification vector stored in \code{GPout}.}

\item{Nindex}{A vector of sequence indices. If set to NULL (default), then the the first sequence in each class labels is considered.}

\item{hcutoff}{A scalar indicating where a horizontal line should be drawn}

\item{show.singletons}{A logical indicating if singleton clusters should be plotted (default=\code{TRUE})}

\item{savefile}{A character vector.  If set to \code{NULL} (default), the figures are not saved and ploted in R, otherwise the figure will be saved to \code{<savefile>Obs<s>Cluster<g>.pdf}, where \code{s} is the sequence index and \code{g} denotes the cluster number.}
}
\description{
Plots the sorted pairwise distances and indicates the position of the largest gap.
}
\details{
For each class label ($g = 1, \dots, G$), a side-by-side boxplot is created.
}

