% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GapProcedure.R
\name{GapProcedure}
\alias{GapProcedure}
\title{The Gap Procedure}
\usage{
GapProcedure(x, submod = "aK80", distance.matrix = NULL,
  outlier.adj = 0.9)
}
\arguments{
\item{x}{Aligned DNA sequence data in \code{DNAbin} (matrix) form.}

\item{submod}{Specifies the evolutionary model to be used for calculating the pairwise genetic
distances between the sequences stored in \code{x}.  Options are: the "aJC", "aK80".}

\item{distance.matrix}{A pre-specified distance matrix used to determine the clusters within \code{x}.}

\item{outlier.adj}{Restricts the search for the larger gap to the first (100*\code{outlier.adj})\% of sorted pairwise distances.}
}
\value{
\describe{
  \item{dist}{A pairwise distance matrix calculated accorinding to \code{submod}.  Returns \code{distance.matrix} if this argument was specified.}
  \item{classification}{A vector indicating the clusters of DNA sequences found by the Gap Procedure.}
  \item{AlgorithmOutputs}{Supplementary outputs (e.g. matrices, CPU times and distances between gaps)
  calculated by the Gap Procedure algorithm.}
}
}
\description{
This function implements the Gap Procedure algorithm described in Vrbik et al.
for aligned DNA sequence data. The partition of the data is based on gaps in pairwise distances.
}
\details{
This function performs the Gap Procedure algorithm as described in Vrbik et al.  The input sequences must be aligned DNA sequences in \code{DNAbin} format.   If no distance matrix is specified in \code{distance.matrix}, this function computes a matrix of pairwise distances an ajusted version of the Kimura 1980 model of DNA evolution.  This adjustment takes into account ambiguous IUPAC codes. For details on these calculations see the Supplmentary Information in Vrbik et al.
}
\examples{
library(ape)
data(woodmouse)
test <- GapProcedure(woodmouse,submod="aK80",outlier.adj=1)

# to see the classifcaiton vector:
test$classification

# to return the pairwise distance matrix:
test$dist
}
\references{
Irene Vrbik, David A Stephens, Michel Roger and Bluma G Brenner. The Gap Procedure: for the identification of phylogenetic clusters in HIV-1 sequence data.s
}
\seealso{
\code{\link{ajc69},\link{ak80}}
}

