% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{lu}
\alias{lu}
\title{Returns the length of the unique elements in a vector x}
\usage{
lu(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An interger indicating the number of unique elements in x
}
\description{
Returns the length of the unique elements in a vector x
}

