% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map2cluster.R
\name{map2cluster}
\alias{map2cluster}
\title{Maps a probabilty matrix (in which each column sums to 1) to classification vector.}
\usage{
map2cluster(z, warn = TRUE, random = FALSE)
}
\arguments{
\item{z}{An $M$ by $N$ probability matrix.  Each column must sum to 1.}

\item{warn}{A logical indicating whether or not a warning should be displayed if any column indexes do not appear in the returned classification vector.}
}
\value{
An integer vector of length $N$.  The $i$th element provides the row index ($1,2,\dots, M$) of the maximum value of the $i$th column of \code{z}.
}
\description{
Maps a probabilty matrix (in which each column sums to 1) to classification vector.
}
\details{
Ties are resolved by randomly sampling a single number from the indices which attain a maximum.
}
\examples{
set.seed(1234)
(ex.matrix <- matrix(sample(100,12), ncol=4, nrow=3))
(prob.matrix <- sweep(ex.matrix,2,colSums(ex.matrix),"/"))
(mapclass <- map2cluster(prob.matrix))

# Create a new rows corresponding to the maximum value of each column
(ex.matrix2 <- rbind(ex.matrix, ex.matrix[cbind(mapclass,1:ncol(ex.matrix))]))
(prob.matrix2 <- sweep(ex.matrix2,2,colSums(ex.matrix2),"/"))

# in the presence of a tie, sequences are assigned to the largest cluster
(map2cluster(prob.matrix2), random=TRUE)
(map2cluster(prob.matrix2), random=TRUE)
}

