% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix2vecInd.R
\name{matrix2vecInd}
\alias{matrix2vecInd}
\title{Converts the matrix index to a vector index}
\usage{
matrix2vecInd(row.ind, col.ind, mat.dim)
}
\arguments{
\item{row.ind}{A integer indicating the row index of a matrix}

\item{col.ind}{An integer indiciation the column index of a matrix}

\item{mat.dim}{A vector of length 2 providing the dimensions of the matrix}
}
\value{
out An interger indicating the scalar index of the matrix corresponding to row index `row.ind` and column index `col.ind`.
}
\description{
Converts the matrix index to a vector index
}
\examples{
(mat <- matrix(1:25, nrow=5))
matrix2vecInd(1,3,dim(mat))

(mat2 <- matrix(1:30, nrow=5))
matrix2vecInd(1,3,dim(mat2))

(mat3 <- matrix(1:30, ncol=5))
matrix2vecInd(1,3,dim(mat3))

matrix2vecInd(c(1,4,6),c(3,1,5),dim(mat3))
}

