% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pifreq}
\alias{pifreq}
\title{Calculates the empirical base frequencies}
\usage{
pifreq(x)
}
\arguments{
\item{x}{A character vector of the sequence data.}
}
\description{
Calculates the empirical base frequencies
}

