% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{renameClusters}
\alias{renameClusters}
\title{Replaces singletons (i.e., elements with value 0) with a unqiue positive valued class labels}
\usage{
renameClusters(class.vec)
}
\arguments{
\item{class.vec}{A vector of class labels}
}
\value{
A new (or identical) class label with stictly positive class labels.
}
\description{
Replaces singletons (i.e., elements with value 0) with a unqiue positive valued class labels
}

