% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmsingletons.R
\name{rmsingletons}
\alias{rmsingletons}
\title{Maps singleton class labels to the class label \code{0}.}
\usage{
rmsingletons(cvec)
}
\arguments{
\item{x}{A vector of class labels.}
}
\value{
A new (or identical) vector of class lables.
}
\description{
Maps singleton class labels to the class label \code{0}.
}
\details{
If there are no singleton (i.e., one member) clusters \code{rmsingletons} will return a vector identical to \code{x}.  Otherwise, all singleton class labels are replaced by zero.
}
\examples{
rmsingletons(c(4,4,8,1,1,1,3,3,6))
rmsingletons(c(4,4,1,1,1,3,3,2,2))
}

