% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqgen2DNAbin.R
\name{seqgen2DNAbin}
\alias{seqgen2DNAbin}
\title{Converts the outputed sequences from \code{seq.gen} to `DNAbin` format}
\usage{
seqgen2DNAbin(read.seqgen.output)
}
\arguments{
\item{read.seqgen.output}{The output from running seq.gen}
}
\value{
The input sequences in DNAbin format
}
\description{
Converts the outputed sequences from \code{seq.gen} to `DNAbin` format
}

