% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GapProcedure.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{A four group simulation.}
\format{A data frame with 100 sequences (rows) and 801 columns.  The first column \code{X.class} corresponds to the class label and the remaining columns (\code{X1}, ... \code{X800}) give the DNA nucleotide code for the corresponding site.}
\usage{
data(simulation)
}
\description{
A dataset containing the class of 100 simulated DNA sequences of length 800.
}
\keyword{datasets}

