% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeHeight.R
\name{treeHeight}
\alias{treeHeight}
\title{Determines the height of a rooted tree}
\usage{
treeHeight(rooted.tree)
}
\arguments{
\item{rooted.tree}{A rooted tree in Newick format.}
}
\value{
The height of a the inputted \code{rooted.tree}
}
\description{
Determines the height of a rooted tree
}

